Starting http server with ss.http.middleware manually and passing to ss.start(server) is not working.
Rpc call responders do not get called/bound. This is because of moving the loading of the responders from the socketstream getServer getter to the load-api task in https://github.com/socketstream/socketstream/pull/590 In this case, the responders are not defined on the server object when loading the WebSockets layer. To fix this, the call to stream(plan.httpServer) should happen after the tasks (specifically after the load-api task) have run. And, the decision to include the start-server task in the load-socketstream default task should be based on plan.httpServer instead of ss.server.httpServer.